code
Agda translation of my original Lean code, with the addition of a
definition of reduction.

I'm not quite happy with the structure of the files yet, but I'm saving
my work for the moment.

I also need to define the correct notion of definitional equality, at
the moment what I call `DefEq` is actually a notion of reduction.